+++ nasa_edx_extensions/__init__.py
@@ -5,4 +5,4 @@ Init module for nasa_edx_extensions.
 from __future__ import unicode_literals
 
 
-__version__ = '1.4.4'
+__version__ = '1.5.4'